tf-keras-efficient-net/requirements.txt
pip==24.3.1
numpy==1.26.4
fastapi==0.109.2
pydantic==2.9.2
tensorflow==2.17.1
keras==3.5.0
uvicorn
pillow
python-multipart
